restructure Go wrapper for distribution
