+++ languages/go/bitwarden_client.go
@@ -3,7 +3,7 @@ package sdk
 import (
 	"encoding/json"
 
-	"github.com/bitwarden/sdk/languages/go/internal/cinterface"
+	"github.com/bitwarden/sm-sdk-go/cinterface"
 )
 
 type BitwardenClient struct {
