+++ languages/go/go.mod
@@ -1,5 +1,8 @@
-module github.com/bitwarden/sdk/languages/go
+module github.com/bitwarden/sm-sdk-go
 
 go 1.18
 
 require github.com/gofrs/uuid v4.4.0+incompatible
+
+replace github.com/bitwarden/sm-sdk-go => ./go
+replace github.com/bitwarden/sm-sdk-go/cinterface => ./go/cinterface
